refactor(scheduler): split scheduler.py god-file into scheduler/ package (#633)
agentwire/scheduler.py (2044 lines) -> agentwire/scheduler/ with cohesive
submodules: models (dataclasses/constants/parsing), state (board load/save
+ backups), schedule (eligibility math + picking + validation), gates
(precondition checks), dispatch (ensure/worktree+PR execution), report
(events/live-state/portal/display), loop (daemon).

Pure restructuring, zero behavior change. The package __init__ re-exports
every symbol at its historical agentwire.scheduler path, and submodules
resolve cross-cutting calls (save_board, _check_gate, _log_event,
_sched_config, datetime, ...) through the package namespace at call time,
so every existing patch()/monkeypatch target in tests keeps working.

Closes #633

Built by dotdev.dev
